requirements.txt
tendermint-abci
pyshamir
phe
requests
click